Work in progress

+++ Tech Cube 9x9x9/software/Infinity Display/src/core/Display.h
@@ -19,7 +19,7 @@ class Display {
  private:
   // Special display effect for all animations
   static uint8_t motionBlur;
-  static uint8_t brightness;
+  static float brightness;
 
  public:
   // Do not use a class contructor to start the display (Arduino compatibility)
@@ -32,7 +32,7 @@ class Display {
   static void setMotionBlur(const uint8_t);
   static uint8_t getMotionBlur();
   // Set the master display brightness value
-  static void setBrightness(const uint8_t);
-  static uint8_t getBrightness();
+  static void setBrightness(const float);
+  static float getBrightness();
 };
 #endif
